Keyset pagination was added to Gitlab a while ago and as of Gitlab v17.0 'users' endpoint only uses keyset pagination.
This adds support for keyset pagination.

As keyset pagination returns the next page as a link in the
in the HTTP response header. RESTClient.pm was changed to return
http headers when called in list mode.

Paginator.pm checks for the pagination='keyset' parameter or 'users'
method and passes keyset parameters to GitLab API method calls.
It then parses the response header for the Link header, pulling
out the returned URL parameters. Storing them for the next call to
next_page.